use alternative path with appliance for RHEL 8 on non-fedora systems
The RHEL 8 seems to be using XFS filysystem that is not mountable
by the appliance generated in RHEL/CentOS 7.5 and therefore we cannot
apply the hacks. Change here add detection that we run on non-Fedora
system and in such case use a separate appliance in /var/tmp.
TODO: document this better once confirmed to run on other systems well.